requirements.txt
# Runtime dependencies for PGVectorRAGIndexer
# For development tools (pytest, black, ruff, mypy), also install requirements-dev.txt

# MCP Server
mcp>=1.0.0

# Core dependencies
psycopg2-binary>=2.9.9
pgvector>=0.2.4
python-dotenv>=1.0.0
watchdog>=3.0.0
xxhash>=3.4.1

# Configuration and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Document processing
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.1.0
unstructured>=0.11.0

# OCR support (scanned documents and images)
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# Embeddings
sentence-transformers>=2.2.2
torch>=2.1.0

# API
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6

# Database migrations
alembic>=1.13.0
sqlalchemy>=2.0.0

# License key validation
PyJWT>=2.8.0